+++ packages/webview_flutter/webview_flutter_wkwebview/test/webkit_webview_controller_test.mocks.dart
@@ -245,7 +245,7 @@ class MockUIScrollView extends _i1.Mock implements _i2.UIScrollView {
           as _i2.UIScrollView);
 
   @override
-  _i3.Future<void> setBackgroundColor(int? value) =>
+  _i3.Future<void> setBackgroundColor(_i2.UIColor? value) =>
       (super.noSuchMethod(
             Invocation.method(#setBackgroundColor, [value]),
             returnValue: _i3.Future<void>.value(),
@@ -559,6 +559,17 @@ class MockWKPreferences extends _i1.Mock implements _i2.WKPreferences {
   _i3.Future<void> setJavaScriptEnabled(bool? enabled) =>
       (super.noSuchMethod(
             Invocation.method(#setJavaScriptEnabled, [enabled]),
+            returnValue: _i3.Future<void>.value(),
+            returnValueForMissingStub: _i3.Future<void>.value(),
+          )
+          as _i3.Future<void>);
+
+  @override
+  _i3.Future<void> setJavaScriptCanOpenWindowsAutomatically(bool? enabled) =>
+      (super.noSuchMethod(
+            Invocation.method(#setJavaScriptCanOpenWindowsAutomatically, [
+              enabled,
+            ]),
             returnValue: _i3.Future<void>.value(),
             returnValueForMissingStub: _i3.Future<void>.value(),
           )
@@ -1480,7 +1491,7 @@ class MockUIViewWKWebView extends _i1.Mock implements _i2.UIViewWKWebView {
           as _i2.UIViewWKWebView);
 
   @override
-  _i3.Future<void> setBackgroundColor(int? value) =>
+  _i3.Future<void> setBackgroundColor(_i2.UIColor? value) =>
       (super.noSuchMethod(
             Invocation.method(#setBackgroundColor, [value]),
             returnValue: _i3.Future<void>.value(),